% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidstats-package.R
\docType{package}
\name{covidstats-package}
\alias{covidstats}
\alias{covidstats-package}
\title{covidstats: A COVID-19 data extraction.}
\description{
A tool for generating RMS office locations COVID-19 metrics.
}
\author{
\strong{Maintainer}: Brice Jabo \email{brice.jabo@rms.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
