% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-get_data.R
\name{get_country_counts}
\alias{get_country_counts}
\title{Retrieve daily COVID-19 data}
\usage{
get_country_counts(country_name)
}
\arguments{
\item{country_name}{ISO3 country code}
}
\value{
A data frame
}
\description{
The get_country_counts function is used to retrieve country
daily COVID-19 case and death counts.
}
\examples{
get_country_counts("CHE")
get_counts(deaths,deaths_url, c("USA"))
}
