% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-get_data.R
\name{get_country_vax}
\alias{get_country_vax}
\title{Retrieve daily COVID-19 vaccination rates}
\usage{
get_country_vax(country_name)
}
\arguments{
\item{country_name}{ISO3 country name}
}
\value{
A data frame
}
\description{
The get_country_vax function is used to retrieve country
daily COVID-19 vaccination rate.
}
\examples{
get_country_vax("CHE")
}
