% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-get_data.R
\name{get_london_counts}
\alias{get_london_counts}
\title{Retrieve London COVID-19 data}
\usage{
get_london_counts()
}
\value{
A data frame
}
\description{
The get_london_counts function is used to generate London daily
COVID-19 case and death counts.
}
\examples{
get_london_counts()
}
