% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-get_data.R
\name{get_london_vx_demog}
\alias{get_london_vx_demog}
\title{Retrieve London vaccination rates}
\usage{
get_london_vx_demog()
}
\value{
A data frame
}
\description{
The get_london_vx_demog is used to retrieve London daily
COVID-19 vaccination rate by age.
}
\examples{
get_london_vx_demog()
}
