% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-get_data.R
\name{get_us_counts}
\alias{get_us_counts}
\title{Retrieve country COVID-19 data}
\usage{
get_us_counts()
}
\value{
A data frame
}
\description{
The get_us_counts function is used to retrieve US county
daily COVID-19 case and death counts.
}
\examples{
get_us_counts(cases)
}
