% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-get_data.R
\name{get_us_msa}
\alias{get_us_msa}
\title{Retrieve US Metropolitan statistical area(MSA) table}
\usage{
get_us_msa()
}
\value{
US MSA table
}
\description{
The get_us_msa function is used to retrieve the US Census
2020 MSA table.
}
\examples{
get_us_msa()
}
