% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-get_data.R
\name{get_us_vax}
\alias{get_us_vax}
\title{Retrieve US state vaccination rates}
\usage{
get_us_vax()
}
\value{
A data frame
}
\description{
The get_us_vax is used to retrieve CDC US state daily COVID-19
vaccination rates.
}
\examples{
get_us_vax()
}
