% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-get_data.R
\name{get_us_vx_demog}
\alias{get_us_vx_demog}
\title{Retrieve US state and demographic vaccination rates}
\usage{
get_us_vx_demog()
}
\value{
A data frame
}
\description{
The get_us_vx_demog function retrieves USFACTS state vaccination
rate by age.
}
\examples{
get_us_vx_demog()
}
