% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-get_data.R
\name{get_zurich_counts}
\alias{get_zurich_counts}
\title{Retrieve Zurich COVID-19 data.}
\usage{
get_zurich_counts()
}
\value{
A data frame
}
\description{
The get_zurich_counts function is used to generate Zurich canton
daily COVID-19 case and death counts.
}
\examples{
get_zurich_counts()
}
