% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-get_data.R
\name{set_country_name}
\alias{set_country_name}
\title{Translate country ISO3 code to country name}
\usage{
set_country_name(country_name)
}
\arguments{
\item{country_name}{ISO3 country code}
}
\value{
A data frame
}
\description{
The set_country_name function is used to translate ISO3 country
codes to country name. Country names are necessary sub-setting CSSE
data by country.
}
\examples{
set_country_name("CHE")
}
